+++ packages/client/src/pages/user/index.photos.vue
@@ -90,7 +90,7 @@ export default defineComponent({
 		> .img {
 			height: 128px;
 			border-radius: 6px;
-			overflow: clip;
+			overflow: hidden; overflow: clip;
 		}
 	}