+++ packages/client/src/ui/_common_/statusbar-rss.vue
@@ -79,9 +79,9 @@ useInterval(tick, Math.max(5000, props.refreshIntervalSec * 1000), {
 			display: inline-block;
 			width: 0.5px;
 			height: var(--height);
-			margin: 0 1em;
+			margin: 0 2em;
 			background: currentColor;
-			opacity: 0.7;
+			opacity: 0.3;
 		}
 	}
 }