+++ packages/client/src/widgets/index.ts
@@ -6,7 +6,7 @@ export default function(app: App) {
 	app.component('MkwTimeline', defineAsyncComponent(() => import('./timeline.vue')));
 	app.component('MkwCalendar', defineAsyncComponent(() => import('./calendar.vue')));
 	app.component('MkwRss', defineAsyncComponent(() => import('./rss.vue')));
-	app.component('MkwRssMarquee', defineAsyncComponent(() => import('./rss-marquee.vue')));
+	app.component('MkwRssTicker', defineAsyncComponent(() => import('./rss-ticker.vue')));
 	app.component('MkwTrends', defineAsyncComponent(() => import('./trends.vue')));
 	app.component('MkwClock', defineAsyncComponent(() => import('./clock.vue')));
 	app.component('MkwActivity', defineAsyncComponent(() => import('./activity.vue')));
@@ -30,7 +30,7 @@ export const widgets = [
 	'timeline',
 	'calendar',
 	'rss',
-	'rssMarquee',
+	'rssTicker',
 	'trends',
 	'clock',
 	'activity',